Extract ResizeAxisPicker and rename to CuboidResizeInteraction
The interaction class was doing two jobs: pick which axis the gesture
controls + bind each finger to a face (one-time decision at begin),
and run per-frame kinematics. The first half is now a pure
ResizeAxisPicker, leaving the interaction class responsible only for
lifecycle and per-frame math.

The class operates purely on cuboid geometry, with no notion of
parcel — the rename matches the neighbours (ARCuboidEntity,
TwoFingerCuboidGesture).
